edit.rs を分割した。
